Finsberg/major refactor (#194)
* Some structural changes - and make tests pass

* Ruff

* Fix mypy

* Fix mode in write_mesh for adios2
