+++ src/adios4dolfinx/__init__.py
@@ -8,6 +8,7 @@
 
 from importlib.metadata import metadata
 
+from .backends import FileMode, get_backend
 from .checkpointing import (
     read_attributes,
     read_function,
@@ -19,7 +20,6 @@ from .checkpointing import (
     write_mesh,
     write_meshtags,
 )
-from .interface import FileMode, get_backend
 from .legacy_readers import read_function_from_legacy_h5, read_mesh_from_legacy_h5
 from .original_checkpoint import write_function_on_input_mesh, write_mesh_input_order
 from .snapshot import snapshot_checkpoint
