+++ src/adios4dolfinx/backends/h5py/__init__.py
@@ -1,0 +1,3 @@
+from . import backend
+
+__all__ = ["backend"]
